Extract MonospaceFontDialog into its own module
Move the self-contained MonospaceFontDialog (an Adw.Window for picking a
monospace terminal font) verbatim out of preferences.py into a new leaf
module sshpilot/monospace_font_dialog.py. The dialog depends only on
GTK/Pango with no preferences state, so it stands alone; preferences.py
imports it back (used by the font-selection action) and keeps it available
for back-compat.

Drop the now-unused PangoFT2 from preferences.py's gi import and add the
proper gi.require_version('PangoFT2', '1.0') in the new module to avoid a
PyGIWarning. preferences.py: 5013 -> 4753 lines.

Adds tests/test_monospace_font_dialog.py guarding the extraction contract
(importable from the new module, re-exported from preferences, public
method surface intact).
